--- src/context/Repo/RepoContext.ts
@@ -1,4 +1,0 @@
-import {createContext} from 'react';
-import {RepoContextType} from './RepoProvider';
-
-export const RepoContext = createContext({} as RepoContextType);
