+++ src/models/Firebase.ts
@@ -1,0 +1,9 @@
+export type FirebaseKey = string;
+
+export interface FirebaseTermsRawResponse {
+  [key: FirebaseKey]: {
+    id: string,
+    term: string,
+    definition: string
+  }
+}
